requirements.txt
Django==5.2.2
mysqlclient==2.2.0
Pillow==10.0.0
reportlab==4.0.4
xlsxwriter==3.1.2
django-crispy-forms==2.0
django-chartjs==2.3.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-barcode==0.15.1
qrcode==7.4.2
django-extensions==3.2.3
django-import-export==3.3.1
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.5.0
python-decouple==3.8
django-cors-headers==4.3.0
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0